[test] split tracer related tests out of test_jit
test_jit is becoming huge again, which makes editor hard to load and
write new tests, this split out the tracer related tests.

[ghstack-poisoned]
